+++ lib/factory_sloth/code_mod.rb
@@ -19,13 +19,16 @@ class FactorySloth::CodeMod
     # rspec call. However, this would make it impossible to use `--fail-fast`,
     # and might make examples fail that are not as idempotent as they should be.
     self.changed_create_calls =
-      create_calls
-      .sort_by { |call| [-call.line, -call.column] }
-      .select { |call| try_patch(call, 'build') || try_patch(call, 'build_stubbed') }
+      create_calls.sort_by { |call| [-call.line, -call.column] }.select do |call|
+        build_result = try_patch(call, 'build')
+        next if build_result == ABORT
+
+        build_result == SUCCESS || try_patch(call, 'build_stubbed') == SUCCESS
+      end
 
     # validate whole spec after changes, e.g. to detect side-effects
     self.ok = changed_create_calls.none? ||
-      FactorySloth::SpecRunner.call(path, patched_code)
+      FactorySloth::SpecRunner.call(path, patched_code).success?
     changed_create_calls.clear unless ok?
     patched_code.replace(original_code) unless ok?
   end
@@ -56,30 +59,56 @@ class FactorySloth::CodeMod
       /\A(?:.*\n){#{call.line - 1}}.{#{call.column}}\K#{call.name}/,
       variant
     )
-    checked_patched_code = with_execution_check(new_patched_code, call.line, variant)
-    if FactorySloth::SpecRunner.call(path, checked_patched_code, line: call.line)
+    checked_patched_code = new_patched_code + checks(call.line, variant)
+
+    result = FactorySloth::SpecRunner.call(path, checked_patched_code, line: call.line)
+    if result.success?
       puts "- #{call.name} in line #{call.line} can be replaced with #{variant}"
       self.patched_code = new_patched_code
+      SUCCESS
+    elsif result.exitstatus == FACTORY_UNUSED_CODE
+      puts "- #{call.name} in line #{call.line} is never executed, skipping"
+      ABORT
+    elsif result.exitstatus == FACTORY_PERSISTED_LATER_CODE
+      ABORT
     end
   end
 
-  def with_execution_check(spec_code, line, variant)
-    spec_code + <<~RUBY
+  ABORT = :ABORT # returned if there is no need to try other variants
+  SUCCESS = :SUCCESS
+
+  FACTORY_UNUSED_CODE          = 77
+  FACTORY_PERSISTED_LATER_CODE = 78
+
+  # This adds code that makes a spec run fail and thus prevents changes if:
+  # a) the patched factory in the given line is never called
+  # b) the built record was persisted later anyway
+  # The rationale behind a) is that things like skipped examples should not
+  # be broken. The rationale behind b) is that not much DB work would be saved,
+  # but diff noise would be increased and ease of editing the example reduced.
+  def checks(line, variant)
+    <<~RUBY
       ; defined?(FactoryBot) && defined?(RSpec) && RSpec.configure do |config|
-        executed_lines = []
+        records_by_line = {} # track records initialized through factories per line
 
         FactoryBot::Syntax::Methods.class_eval do
-          alias ___original_#{variant} #{variant}
+          alias ___original_#{variant} #{variant} # e.g. ___original_build build
 
-          define_method("#{variant}") do |*args, **kwargs, &blk|
-            executed_lines << caller_locations(1, 1)&.first&.lineno
-            ___original_#{variant}(*args, **kwargs, &blk)
+          define_method("#{variant}") do |*args, **kwargs, &blk| # e.g. build
+            result = ___original_#{variant}(*args, **kwargs, &blk)
+            list = records_by_line[caller_locations(1, 1)&.first&.lineno] ||= []
+            list.concat([result].flatten) # to work with single, list, and pair
+            result
           end
         end
 
         config.after(:suite) do
-          executed_lines.include?(#{line}) ||
-            fail("unused factory in line #{line} - will not be modified")
+          records = records_by_line[#{line}]
+          records&.any? || exit!(#{FACTORY_UNUSED_CODE})
+          unless "#{variant}".include?('stub') # factory_bot stub stubs persisted? as true
+            records.any? { |r| r.respond_to?(:persisted?) && r.persisted? } &&
+              exit!(#{FACTORY_PERSISTED_LATER_CODE})
+          end
         end
       end
     RUBY