+++ spec/factory_bot_stub.rb
@@ -2,9 +2,9 @@ module FactoryBot
   module Syntax
     module Methods
       %w[build build_stubbed create].each do |variant|
-        define_method(variant) { |*| true }
-        define_method("#{variant}_list") { |_name, n, *| n.times.map { true } }
-        define_method("#{variant}_pair") { |_name, n, *| [true, true] }
+        define_method(variant) { |name, *| Record.new(variant: variant, name: name) }
+        define_method("#{variant}_list") { |name, n, *| (1..n).map { send(variant, name) } }
+        define_method("#{variant}_pair") { |name, *| (1..2).map { send(variant, name) } }
       end
     end
   end
@@ -12,3 +12,14 @@ module FactoryBot
 end
 
 include FactoryBot::Syntax::Methods
+
+Record = Struct.new(:name, :variant, keyword_init: true) do
+  def save!
+    raise 'stubbed, cant save!' if variant =~ /stub/
+    @persisted = true
+  end
+
+  def persisted?
+    !!@persisted
+  end
+end